pyproject.toml
[tool.poetry]
name = "bjj-journey"
version = "0.1.0"
description = ""
authors = ["dsli9 <liderek3@gmail.com>"]
readme = "README.md"
packages = [{include = "bjj_journey", from = "src"}]

[tool.poetry.dependencies]
python = "~3.8"
alembic = "^1.10.2"
psycopg2-binary = "^2.9.5"
gspread = "^5.8.0"
python-dotenv = "^1.0.0"
streamlit = "^1.21.0"
dbt-postgres = "^1.4.5"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
mypy = "^1.2.0"
pandas-stubs = "^1.5.3.230321"
pylint = "^2.17.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.mypy.overrides]]
module = "gspread.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "altair.*"
ignore_missing_imports = true

[tool.pylint.format]
max-line-length = 88
